# web/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Kimi Code CLI Web Interface", "version": "0.1.0" },
  "paths": {
    "/api/config/": {
      "get": {
        "tags": ["config"],
        "summary": "Get global (novel-cli) config snapshot",
        "description": "Get global (novel-cli) config snapshot.",
        "operationId": "get_global_config_api_config__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GlobalConfig" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["config"],
        "summary": "Update global (novel-cli) default model/thinking",
        "description": "Update global (novel-cli) default model/thinking.",
        "operationId": "update_global_config_api_config__patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGlobalConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateGlobalConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/config/toml": {
      "get": {
        "tags": ["config"],
        "summary": "Get novel-cli config.toml",
        "description": "Get novel-cli config.toml.",
        "operationId": "get_config_toml_api_config_toml_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ConfigToml" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["config"],
        "summary": "Update novel-cli config.toml",
        "description": "Update novel-cli config.toml.",
        "operationId": "update_config_toml_api_config_toml_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigTomlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateConfigTomlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigModel": {
        "properties": {
          "provider": { "type": "string", "title": "Provider" },
          "model": { "type": "string", "title": "Model" },
          "max_context_size": {
            "type": "integer",
            "title": "Max Context Size"
          },
          "capabilities": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/ModelCapability" },
                "type": "array",
                "uniqueItems": true
              },
              { "type": "null" }
            ],
            "title": "Capabilities"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Model key in novel-cli config (Config.models)"
          },
          "provider_type": {
            "$ref": "#/components/schemas/ProviderType",
            "description": "Provider type (LLMProvider.type)"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "max_context_size",
          "name",
          "provider_type"
        ],
        "title": "ConfigModel",
        "description": "Model configuration for frontend."
      },
      "ConfigToml": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Raw TOML content"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Path to config file"
          }
        },
        "type": "object",
        "required": ["content", "path"],
        "title": "ConfigToml",
        "description": "Raw config.toml content."
      },
...
      "GlobalConfig": {
        "properties": {
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "description": "Current default model key"
          },
          "default_thinking": {
            "type": "boolean",
            "title": "Default Thinking",
            "description": "Current default thinking mode"
          },
          "models": {
            "items": { "$ref": "#/components/schemas/ConfigModel" },
            "type": "array",
            "title": "Models",
            "description": "All configured models"
          }
        },
        "type": "object",
        "required": ["default_model", "default_thinking", "models"],
        "title": "GlobalConfig",
        "description": "Global configuration snapshot for frontend."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelCapability": {
        "type": "string",
        "enum": ["image_in", "video_in", "thinking", "always_thinking"]
      },
...
      "ProviderType": {
        "type": "string",
        "enum": [
          "kimi",
          "openai_legacy",
          "openai_responses",
          "anthropic",
          "google_genai",
          "gemini",
          "vertexai",
          "_echo",
          "_scripted_echo",
          "_chaos"
        ]
      },
...
      "UpdateConfigTomlRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "New TOML content"
          }
        },
        "type": "object",
        "required": ["content"],
        "title": "UpdateConfigTomlRequest",
        "description": "Request to update config.toml."
      },
      "UpdateConfigTomlResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the update was successful"
          },
          "error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "type": "object",
        "required": ["success"],
        "title": "UpdateConfigTomlResponse",
        "description": "Response after updating config.toml."
      },
      "UpdateGlobalConfigRequest": {
        "properties": {
          "default_model": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Default Model",
            "description": "New default model key"
          },
          "default_thinking": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Default Thinking",
            "description": "New default thinking mode"
          },
          "restart_running_sessions": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Restart Running Sessions",
            "description": "Whether to restart running sessions"
          },
          "force_restart_busy_sessions": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Force Restart Busy Sessions",
            "description": "Whether to force restart busy sessions"
          }
        },
        "type": "object",
        "title": "UpdateGlobalConfigRequest",
        "description": "Request to update global config."
      },
      "UpdateGlobalConfigResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/GlobalConfig",
            "description": "Updated config snapshot"
          },
          "restarted_session_ids": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Restarted Session Ids",
            "description": "IDs of restarted sessions"
          },
          "skipped_busy_session_ids": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Skipped Busy Session Ids",
            "description": "IDs of busy sessions that were skipped"
          }
        },
        "type": "object",
        "required": ["config"],
        "title": "UpdateGlobalConfigResponse",
        "description": "Response after updating global config."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}